♻️ refactor: isolate local Agent Signal scheduler
