+++ apps/server/src/workflows/agentSignal/impls/index.ts
@@ -1,0 +1,1 @@
+export { scheduleLocalAgentSignalRun } from './local';
